+++ fmtutil/fmtutil_test.go
@@ -128,39 +128,55 @@ velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magna
 
 Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea 
 commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae 
-consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?
-`
-	result := `  Sed ut perspiciatis unde omnis iste 
-  natus error sit voluptatem accusantium 
-  doloremque laudantium, totam rem 
-  aperiam, eaque ipsa quae ab illo 
-  inventore veritatis et quasi 
-  architecto beatae vitae dicta sunt 
-  explicabo. Nemo enim ipsam voluptatem 
-  quia voluptas sit aspernatur aut odit 
-  aut fugit, sed quia consequuntur magni 
-  dolores eos qui ratione voluptatem 
-  sequi nesciunt, cum soluta nobis est 
+consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?`
+
+	result := `  Sed ut perspiciatis unde omnis iste
+  natus error sit voluptatem accusantium
+  doloremque laudantium, totam rem
+  aperiam, eaque ipsa quae ab illo
+  inventore veritatis et quasi
+  architecto beatae vitae dicta sunt
+  explicabo. Nemo enim ipsam voluptatem
+  quia voluptas sit aspernatur aut odit
+  aut fugit, sed quia consequuntur magni
+  dolores eos qui ratione voluptatem
+  sequi nesciunt, cum soluta nobis est
   caparet.
 
-  Neque porro quisquam est, qui dolorem 
-  ipsum quia dolor sit amet, 
-  consectetur, adipisci velit, sed quia 
-  non numquam eius modi tempora incidunt 
-  ut labore et dolore magnam aliquam 
+  Neque porro quisquam est, qui dolorem
+  ipsum quia dolor sit amet,
+  consectetur, adipisci velit, sed quia
+  non numquam eius modi tempora incidunt
+  ut labore et dolore magnam aliquam
   quaerat voluptatem.
 
-  Ut enim ad minima veniam, quis nostrum 
-  exercitationem ullam corporis suscipit 
-  laboriosam, nisi ut aliquid ex ea 
-  commodi consequatur? Quis autem vel 
-  eum iure reprehenderit qui in ea 
-  voluptate velit esse quam nihil 
-  molestiae consequatur, vel illum qui 
-  dolorem eum fugiat quo voluptas nulla 
-  pariatur?`
+  Ut enim ad minima veniam, quis nostrum
+  exercitationem ullam corporis suscipit
+  laboriosam, nisi ut aliquid ex ea
+  commodi consequatur? Quis autem vel
+  eum iure reprehenderit qui in ea
+  voluptate velit esse quam nihil
+  molestiae consequatur, vel illum qui
+  dolorem eum fugiat quo voluptas nulla
+  pariatur?
+`
 
 	c.Assert(Wrap(input, "  ", 40), Equals, result)
+
+	input = "Sed ut \033[40;38;5;82mperspiciatis \x1b[30;48;5;82munde\x1B[0m omnis iste natus error sit voluptatem accusantium doloremque"
+	result = "  Sed ut \x1b[40;38;5;82mperspiciatis \x1b[30;48;5;82munde\x1b[0m\n" +
+		"  omnis iste natus error sit\n" +
+		"  voluptatem accusantium\n" +
+		"  doloremque\n"
+
+	c.Assert(Wrap(input, "  ", 30), Equals, result)
+
+	input = "abcd1234abcd1234abcd1234abcd1234 abcd abcd abcd 1234 1234"
+	result = "  abcd1234abcd1234abcd1234abcd1234\n" +
+		"  abcd abcd abcd\n" +
+		"  1234 1234\n"
+
+	c.Assert(Wrap(input, "  ", 20), Equals, result)
 }
 
 func (s *FmtUtilSuite) TestSeparator(c *C) {
@@ -213,5 +229,11 @@ func (s *FmtUtilSuite) BenchmarkPrettyNum(c *C) {
 func (s *FmtUtilSuite) BenchmarkPrettyNumFloat(c *C) {
 	for i := 0; i < c.N; i++ {
 		PrettyNum(-123456.15)
+	}
+}
+
+func (s *FmtUtilSuite) BenchmarkWrap(c *C) {
+	for i := 0; i < c.N; i++ {
+		Wrap("0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0", "  ", 10)
 	}
 }