+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * `[ansi]` Added methods for working with byte slices
 * `[secstr]` Added method `IsEmpty` for checking empty strings
+* `[fmtutil]` Method `Wrap` was rewritten from scratch (`3386 ns/op | 1152 B/op | 40 allocs/op` → `1322 ns/op | 416 B/op | 5 allocs/op`)
+* `[fmtutil]` Method `Wrap` now ignores ANSI/VT100 control sequences
 * `[usage]` Code refactoring
 
 ### 12.39.1